[AMDGPU][LTO] Introduce AMDGPUCloneModuleLDS
The purpose of this pass is to ensure that the
combined module contains as many LDS global variables
as there are kernels that (indirectly) access them.
As LDS variables behave like C++ static variables,
it is important that each partition contains a
unique copy of the variable on a per kernel basis.
This representation also prepares the combined
module to eliminate cross-module dependencies of
LDS variables.

This pass operates as follows:
1. Firstly, traverse the call graph from each kernel
   to determine the number of kernels calling each
   device function.
2. For each LDS global variable GV, determine the
   function F that defines it. Collect it's caller
   functions. Clone F and GV, and finally insert a
   call/invoke instruction in each caller function.

Change-Id: I998291a389ea3db10de9122f08fe55c981da6049

[AMDGPUCloneModuleLDS] Add GlobalAlias support

Also includes some minor nits

Change-Id: I94dd1845d9896e797f25ae7968a0d9b055f89161

[DO NOT LAND] Clever-ish module splitting

Test for smarter module splitting for --lto-partitions

Includes a patch to make LowerModuleLDS work with fullLTO pipeline.

Change-Id: I745014f378f6d3e4a55c2ea521c78fd87e76e362