[LSV] Enable LSV to handle mismatched types
Prior to this patch, LSV built chains by
gathering a collection of loads/stores with
value operands of the same bitwidth of their
respective scalar type. Instead, the LSV may
consider the bitwidth of the value type. This
allows for grouping mismatched types like i32
and <2 x i16> in the same equivalence class.

The above change asserts that each chain element
will have the same total bitwidth. Hence, for
chains with multiple types of elements, LSV
could use intN as the type of the vectorized
load/store, where N is the bitwidth of each
element. Except for the trivial case where each
element are of the same type, consider that
type without overthinking.

TODO: Add more info about the patch...

Change-Id: I987995360321d9b623ea619379466efe829e585c