[InstCombine] Iterative replacement in PtrReplacer
This patch enhances the PtrReplacer as follows:

1. Users are now collected iteratively to be
generous on the stack. In the case of PHIs with
incoming values which have not yet been visited,
they are pushed back into the stack for
reconsideration.

2. Replace users of the pointer root in a
reverse-postorder traversal, instead of a simple
traversal over the collected users. This reordering
ensures that the uses of an instruction are
replaced before replacing the instruction itself.

3. During the replacement of PHI, use the same
incoming value if it does not have a replacement.

This patch specifically fixes the case when an
incoming value of a PHI is addrspacecasted.